+++ src/edutap/data_provider/__init__.py
@@ -2,8 +2,15 @@
 
 from importlib.metadata import version
 
-from .vocabulary import FieldKind, PassLifecycleState, WalletType
+from .vocabulary import FieldKind, HolderState, InstanceState, IssuanceState, WalletType
 
 __version__ = version("edutap.data_provider")
 
-__all__ = ["FieldKind", "PassLifecycleState", "WalletType", "__version__"]
+__all__ = [
+    "FieldKind",
+    "HolderState",
+    "InstanceState",
+    "IssuanceState",
+    "WalletType",
+    "__version__",
+]